Move SPIRVMap initializations out of OCLUtil.h
The map initialization unnecessarily ends up in almost every
translation unit due to transitive inclusion.
