+++ lib/SPIRV/OCLUtil.cpp
@@ -96,11 +96,548 @@ namespace OCLUtil {
 #define SPIRV_IMAGE_ADDR_SPACE SPIRAS_Global
 #endif
 
+} // namespace OCLUtil
+
+///////////////////////////////////////////////////////////////////////////////
+//
+// Map definitions
+//
+///////////////////////////////////////////////////////////////////////////////
+
+using namespace OCLUtil;
+namespace SPIRV {
+
+template <> void SPIRVMap<OCLMemFenceKind, MemorySemanticsMask>::init() {
+  add(OCLMF_Local, MemorySemanticsWorkgroupMemoryMask);
+  add(OCLMF_Global, MemorySemanticsCrossWorkgroupMemoryMask);
+  add(OCLMF_Image, MemorySemanticsImageMemoryMask);
+}
+
+template <>
+void SPIRVMap<OCLMemFenceExtendedKind, MemorySemanticsMask>::init() {
+  add(OCLMFEx_Local, MemorySemanticsWorkgroupMemoryMask);
+  add(OCLMFEx_Global, MemorySemanticsCrossWorkgroupMemoryMask);
+  add(OCLMFEx_Local_Global, MemorySemanticsWorkgroupMemoryMask |
+                                MemorySemanticsCrossWorkgroupMemoryMask);
+  add(OCLMFEx_Image, MemorySemanticsImageMemoryMask);
+  add(OCLMFEx_Image_Local,
+      MemorySemanticsWorkgroupMemoryMask | MemorySemanticsImageMemoryMask);
+  add(OCLMFEx_Image_Global,
+      MemorySemanticsCrossWorkgroupMemoryMask | MemorySemanticsImageMemoryMask);
+  add(OCLMFEx_Image_Local_Global, MemorySemanticsWorkgroupMemoryMask |
+                                      MemorySemanticsCrossWorkgroupMemoryMask |
+                                      MemorySemanticsImageMemoryMask);
+}
+
+template <>
+void SPIRVMap<OCLMemOrderKind, unsigned, MemorySemanticsMask>::init() {
+  add(OCLMO_relaxed, MemorySemanticsMaskNone);
+  add(OCLMO_acquire, MemorySemanticsAcquireMask);
+  add(OCLMO_release, MemorySemanticsReleaseMask);
+  add(OCLMO_acq_rel, MemorySemanticsAcquireReleaseMask);
+  add(OCLMO_seq_cst, MemorySemanticsSequentiallyConsistentMask);
+}
+
+template <> void SPIRVMap<OCLScopeKind, Scope>::init() {
+  add(OCLMS_work_item, ScopeInvocation);
+  add(OCLMS_work_group, ScopeWorkgroup);
+  add(OCLMS_device, ScopeDevice);
+  add(OCLMS_all_svm_devices, ScopeCrossDevice);
+  add(OCLMS_sub_group, ScopeSubgroup);
+}
+
+template <> void SPIRVMap<std::string, SPIRVGroupOperationKind>::init() {
+  add("reduce", GroupOperationReduce);
+  add("scan_inclusive", GroupOperationInclusiveScan);
+  add("scan_exclusive", GroupOperationExclusiveScan);
+  add("ballot_bit_count", GroupOperationReduce);
+  add("ballot_inclusive_scan", GroupOperationInclusiveScan);
+  add("ballot_exclusive_scan", GroupOperationExclusiveScan);
+  add("non_uniform_reduce", GroupOperationReduce);
+  add("non_uniform_scan_inclusive", GroupOperationInclusiveScan);
+  add("non_uniform_scan_exclusive", GroupOperationExclusiveScan);
+  add("non_uniform_reduce_logical", GroupOperationReduce);
+  add("non_uniform_scan_inclusive_logical", GroupOperationInclusiveScan);
+  add("non_uniform_scan_exclusive_logical", GroupOperationExclusiveScan);
+  add("clustered_reduce", GroupOperationClusteredReduce);
+}
+
+template <> void SPIRVMap<std::string, SPIRVFPRoundingModeKind>::init() {
+  add("rte", FPRoundingModeRTE);
+  add("rtz", FPRoundingModeRTZ);
+  add("rtp", FPRoundingModeRTP);
+  add("rtn", FPRoundingModeRTN);
+}
+
+template <> void SPIRVMap<OclExt::Kind, std::string>::init() {
+#define _SPIRV_OP(x) add(OclExt::x, #x);
+  _SPIRV_OP(cl_images)
+  _SPIRV_OP(cl_doubles)
+  _SPIRV_OP(cl_khr_int64_base_atomics)
+  _SPIRV_OP(cl_khr_int64_extended_atomics)
+  _SPIRV_OP(cl_khr_fp16)
+  _SPIRV_OP(cl_khr_gl_sharing)
+  _SPIRV_OP(cl_khr_gl_event)
+  _SPIRV_OP(cl_khr_d3d10_sharing)
+  _SPIRV_OP(cl_khr_media_sharing)
+  _SPIRV_OP(cl_khr_d3d11_sharing)
+  _SPIRV_OP(cl_khr_global_int32_base_atomics)
+  _SPIRV_OP(cl_khr_global_int32_extended_atomics)
+  _SPIRV_OP(cl_khr_local_int32_base_atomics)
+  _SPIRV_OP(cl_khr_local_int32_extended_atomics)
+  _SPIRV_OP(cl_khr_byte_addressable_store)
+  _SPIRV_OP(cl_khr_3d_image_writes)
+  _SPIRV_OP(cl_khr_gl_msaa_sharing)
+  _SPIRV_OP(cl_khr_depth_images)
+  _SPIRV_OP(cl_khr_gl_depth_images)
+  _SPIRV_OP(cl_khr_subgroups)
+  _SPIRV_OP(cl_khr_mipmap_image)
+  _SPIRV_OP(cl_khr_mipmap_image_writes)
+  _SPIRV_OP(cl_khr_egl_event)
+  _SPIRV_OP(cl_khr_srgb_image_writes)
+#undef _SPIRV_OP
+}
+
+template <> void SPIRVMap<OclExt::Kind, SPIRVCapabilityKind>::init() {
+  add(OclExt::cl_images, CapabilityImageBasic);
+  add(OclExt::cl_doubles, CapabilityFloat64);
+  add(OclExt::cl_khr_int64_base_atomics, CapabilityInt64Atomics);
+  add(OclExt::cl_khr_int64_extended_atomics, CapabilityInt64Atomics);
+  add(OclExt::cl_khr_fp16, CapabilityFloat16);
+  add(OclExt::cl_khr_subgroups, CapabilityGroups);
+  add(OclExt::cl_khr_mipmap_image, CapabilityImageMipmap);
+  add(OclExt::cl_khr_mipmap_image_writes, CapabilityImageMipmap);
+}
+
+/// Map OpenCL work functions to SPIR-V builtin variables.
+template <> void SPIRVMap<std::string, SPIRVBuiltinVariableKind>::init() {
+  add("get_work_dim", BuiltInWorkDim);
+  add("get_global_size", BuiltInGlobalSize);
+  add("get_global_id", BuiltInGlobalInvocationId);
+  add("get_global_offset", BuiltInGlobalOffset);
+  add("get_local_size", BuiltInWorkgroupSize);
+  add("get_enqueued_local_size", BuiltInEnqueuedWorkgroupSize);
+  add("get_local_id", BuiltInLocalInvocationId);
+  add("get_num_groups", BuiltInNumWorkgroups);
+  add("get_group_id", BuiltInWorkgroupId);
+  add("get_global_linear_id", BuiltInGlobalLinearId);
+  add("get_local_linear_id", BuiltInLocalInvocationIndex);
+  // cl_khr_subgroups
+  add("get_sub_group_size", BuiltInSubgroupSize);
+  add("get_max_sub_group_size", BuiltInSubgroupMaxSize);
+  add("get_num_sub_groups", BuiltInNumSubgroups);
+  add("get_enqueued_num_sub_groups", BuiltInNumEnqueuedSubgroups);
+  add("get_sub_group_id", BuiltInSubgroupId);
+  add("get_sub_group_local_id", BuiltInSubgroupLocalInvocationId);
+  // cl_khr_subgroup_ballot
+  add("get_sub_group_eq_mask", BuiltInSubgroupEqMask);
+  add("get_sub_group_ge_mask", BuiltInSubgroupGeMask);
+  add("get_sub_group_gt_mask", BuiltInSubgroupGtMask);
+  add("get_sub_group_le_mask", BuiltInSubgroupLeMask);
+  add("get_sub_group_lt_mask", BuiltInSubgroupLtMask);
+}
+
+// Maps uniqued OCL builtin function name to SPIR-V op code.
+// A uniqued OCL builtin function name may be different from the real
+// OCL builtin function name. e.g. instead of atomic_min, atomic_umin
+// is used for atomic_min with unsigned integer parameter.
+// work_group_ and sub_group_ functions are unified as group_ functions
+// except work_group_barrier.
+class SPIRVInstruction;
+template <> void SPIRVMap<std::string, Op, SPIRVInstruction>::init() {
+#define _SPIRV_OP(x, y) add("atom_" #x, OpAtomic##y);
+  // cl_khr_int64_base_atomics builtins
+  _SPIRV_OP(add, IAdd)
+  _SPIRV_OP(sub, ISub)
+  _SPIRV_OP(xchg, Exchange)
+  _SPIRV_OP(dec, IDecrement)
+  _SPIRV_OP(inc, IIncrement)
+  _SPIRV_OP(cmpxchg, CompareExchange)
+  // cl_khr_int64_extended_atomics builtins
+  _SPIRV_OP(min, SMin)
+  _SPIRV_OP(max, SMax)
+  _SPIRV_OP(and, And)
+  _SPIRV_OP(or, Or)
+  _SPIRV_OP(xor, Xor)
+#undef _SPIRV_OP
+#define _SPIRV_OP(x, y) add("atomic_" #x, Op##y);
+  // CL 2.0 atomic builtins
+  _SPIRV_OP(flag_test_and_set_explicit, AtomicFlagTestAndSet)
+  _SPIRV_OP(flag_clear_explicit, AtomicFlagClear)
+  _SPIRV_OP(load_explicit, AtomicLoad)
+  _SPIRV_OP(store_explicit, AtomicStore)
+  _SPIRV_OP(exchange_explicit, AtomicExchange)
+  _SPIRV_OP(compare_exchange_strong_explicit, AtomicCompareExchange)
+  _SPIRV_OP(compare_exchange_weak_explicit, AtomicCompareExchangeWeak)
+  _SPIRV_OP(inc, AtomicIIncrement)
+  _SPIRV_OP(dec, AtomicIDecrement)
+  _SPIRV_OP(fetch_add_explicit, AtomicIAdd)
+  _SPIRV_OP(fetch_sub_explicit, AtomicISub)
+  _SPIRV_OP(fetch_umin_explicit, AtomicUMin)
+  _SPIRV_OP(fetch_umax_explicit, AtomicUMax)
+  _SPIRV_OP(fetch_min_explicit, AtomicSMin)
+  _SPIRV_OP(fetch_max_explicit, AtomicSMax)
+  _SPIRV_OP(fetch_and_explicit, AtomicAnd)
+  _SPIRV_OP(fetch_or_explicit, AtomicOr)
+  _SPIRV_OP(fetch_xor_explicit, AtomicXor)
+#undef _SPIRV_OP
+#define _SPIRV_OP(x, y) add(#x, Op##y);
+  _SPIRV_OP(dot, Dot)
+  _SPIRV_OP(async_work_group_copy, GroupAsyncCopy)
+  _SPIRV_OP(async_work_group_strided_copy, GroupAsyncCopy)
+  _SPIRV_OP(wait_group_events, GroupWaitEvents)
+  _SPIRV_OP(isequal, FOrdEqual)
+  _SPIRV_OP(isnotequal, FUnordNotEqual)
+  _SPIRV_OP(isgreater, FOrdGreaterThan)
+  _SPIRV_OP(isgreaterequal, FOrdGreaterThanEqual)
+  _SPIRV_OP(isless, FOrdLessThan)
+  _SPIRV_OP(islessequal, FOrdLessThanEqual)
+  _SPIRV_OP(islessgreater, LessOrGreater)
+  _SPIRV_OP(isordered, Ordered)
+  _SPIRV_OP(isunordered, Unordered)
+  _SPIRV_OP(isfinite, IsFinite)
+  _SPIRV_OP(isinf, IsInf)
+  _SPIRV_OP(isnan, IsNan)
+  _SPIRV_OP(isnormal, IsNormal)
+  _SPIRV_OP(signbit, SignBitSet)
+  _SPIRV_OP(any, Any)
+  _SPIRV_OP(all, All)
+  _SPIRV_OP(popcount, BitCount)
+  _SPIRV_OP(get_fence, GenericPtrMemSemantics)
+  // CL 2.0 kernel enqueue builtins
+  _SPIRV_OP(enqueue_marker, EnqueueMarker)
+  _SPIRV_OP(enqueue_kernel, EnqueueKernel)
+  _SPIRV_OP(get_kernel_sub_group_count_for_ndrange_impl,
+            GetKernelNDrangeSubGroupCount)
+  _SPIRV_OP(get_kernel_max_sub_group_size_for_ndrange_impl,
+            GetKernelNDrangeMaxSubGroupSize)
+  _SPIRV_OP(get_kernel_work_group_size_impl, GetKernelWorkGroupSize)
+  _SPIRV_OP(get_kernel_preferred_work_group_size_multiple_impl,
+            GetKernelPreferredWorkGroupSizeMultiple)
+  _SPIRV_OP(retain_event, RetainEvent)
+  _SPIRV_OP(release_event, ReleaseEvent)
+  _SPIRV_OP(create_user_event, CreateUserEvent)
+  _SPIRV_OP(is_valid_event, IsValidEvent)
+  _SPIRV_OP(set_user_event_status, SetUserEventStatus)
+  _SPIRV_OP(capture_event_profiling_info, CaptureEventProfilingInfo)
+  _SPIRV_OP(get_default_queue, GetDefaultQueue)
+  _SPIRV_OP(ndrange_1D, BuildNDRange)
+  _SPIRV_OP(ndrange_2D, BuildNDRange)
+  _SPIRV_OP(ndrange_3D, BuildNDRange)
+  // Generic Address Space Casts
+  _SPIRV_OP(to_global, GenericCastToPtrExplicit)
+  _SPIRV_OP(to_local, GenericCastToPtrExplicit)
+  _SPIRV_OP(to_private, GenericCastToPtrExplicit)
+  // CL 2.0 pipe builtins
+  _SPIRV_OP(read_pipe_2, ReadPipe)
+  _SPIRV_OP(write_pipe_2, WritePipe)
+  _SPIRV_OP(read_pipe_2_bl, ReadPipeBlockingINTEL)
+  _SPIRV_OP(write_pipe_2_bl, WritePipeBlockingINTEL)
+  _SPIRV_OP(read_pipe_4, ReservedReadPipe)
+  _SPIRV_OP(write_pipe_4, ReservedWritePipe)
+  _SPIRV_OP(reserve_read_pipe, ReserveReadPipePackets)
+  _SPIRV_OP(reserve_write_pipe, ReserveWritePipePackets)
+  _SPIRV_OP(commit_read_pipe, CommitReadPipe)
+  _SPIRV_OP(commit_write_pipe, CommitWritePipe)
+  _SPIRV_OP(is_valid_reserve_id, IsValidReserveId)
+  _SPIRV_OP(group_reserve_read_pipe, GroupReserveReadPipePackets)
+  _SPIRV_OP(group_reserve_write_pipe, GroupReserveWritePipePackets)
+  _SPIRV_OP(group_commit_read_pipe, GroupCommitReadPipe)
+  _SPIRV_OP(group_commit_write_pipe, GroupCommitWritePipe)
+  _SPIRV_OP(get_pipe_num_packets_ro, GetNumPipePackets)
+  _SPIRV_OP(get_pipe_num_packets_wo, GetNumPipePackets)
+  _SPIRV_OP(get_pipe_max_packets_ro, GetMaxPipePackets)
+  _SPIRV_OP(get_pipe_max_packets_wo, GetMaxPipePackets)
+  // CL 2.0 workgroup builtins
+  _SPIRV_OP(group_all, GroupAll)
+  _SPIRV_OP(group_any, GroupAny)
+  _SPIRV_OP(group_broadcast, GroupBroadcast)
+  _SPIRV_OP(group_iadd, GroupIAdd)
+  _SPIRV_OP(group_fadd, GroupFAdd)
+  _SPIRV_OP(group_fmin, GroupFMin)
+  _SPIRV_OP(group_umin, GroupUMin)
+  _SPIRV_OP(group_smin, GroupSMin)
+  _SPIRV_OP(group_fmax, GroupFMax)
+  _SPIRV_OP(group_umax, GroupUMax)
+  _SPIRV_OP(group_smax, GroupSMax)
+  // CL image builtins
+  _SPIRV_OP(SampledImage, SampledImage)
+  _SPIRV_OP(ImageSampleExplicitLod, ImageSampleExplicitLod)
+  _SPIRV_OP(read_image, ImageRead)
+  _SPIRV_OP(write_image, ImageWrite)
+  _SPIRV_OP(get_image_channel_data_type, ImageQueryFormat)
+  _SPIRV_OP(get_image_channel_order, ImageQueryOrder)
+  _SPIRV_OP(get_image_num_mip_levels, ImageQueryLevels)
+  _SPIRV_OP(get_image_num_samples, ImageQuerySamples)
+  // Intel Subgroups builtins
+  _SPIRV_OP(intel_sub_group_shuffle, SubgroupShuffleINTEL)
+  _SPIRV_OP(intel_sub_group_shuffle_down, SubgroupShuffleDownINTEL)
+  _SPIRV_OP(intel_sub_group_shuffle_up, SubgroupShuffleUpINTEL)
+  _SPIRV_OP(intel_sub_group_shuffle_xor, SubgroupShuffleXorINTEL)
+  // Intel media_block_io builtins
+  _SPIRV_OP(intel_sub_group_media_block_read, SubgroupImageMediaBlockReadINTEL)
+  _SPIRV_OP(intel_sub_group_media_block_write,
+            SubgroupImageMediaBlockWriteINTEL)
+  // cl_khr_subgroup_non_uniform_vote
+  _SPIRV_OP(group_elect, GroupNonUniformElect)
+  _SPIRV_OP(group_non_uniform_all, GroupNonUniformAll)
+  _SPIRV_OP(group_non_uniform_any, GroupNonUniformAny)
+  _SPIRV_OP(group_non_uniform_all_equal, GroupNonUniformAllEqual)
+  // cl_khr_subgroup_ballot
+  _SPIRV_OP(group_non_uniform_broadcast, GroupNonUniformBroadcast)
+  _SPIRV_OP(group_broadcast_first, GroupNonUniformBroadcastFirst)
+  _SPIRV_OP(group_ballot, GroupNonUniformBallot)
+  _SPIRV_OP(group_inverse_ballot, GroupNonUniformInverseBallot)
+  _SPIRV_OP(group_ballot_bit_extract, GroupNonUniformBallotBitExtract)
+  _SPIRV_OP(group_ballot_bit_count_iadd, GroupNonUniformBallotBitCount)
+  _SPIRV_OP(group_ballot_find_lsb, GroupNonUniformBallotFindLSB)
+  _SPIRV_OP(group_ballot_find_msb, GroupNonUniformBallotFindMSB)
+  // cl_khr_subgroup_non_uniform_arithmetic
+  _SPIRV_OP(group_non_uniform_iadd, GroupNonUniformIAdd)
+  _SPIRV_OP(group_non_uniform_fadd, GroupNonUniformFAdd)
+  _SPIRV_OP(group_non_uniform_imul, GroupNonUniformIMul)
+  _SPIRV_OP(group_non_uniform_fmul, GroupNonUniformFMul)
+  _SPIRV_OP(group_non_uniform_smin, GroupNonUniformSMin)
+  _SPIRV_OP(group_non_uniform_umin, GroupNonUniformUMin)
+  _SPIRV_OP(group_non_uniform_fmin, GroupNonUniformFMin)
+  _SPIRV_OP(group_non_uniform_smax, GroupNonUniformSMax)
+  _SPIRV_OP(group_non_uniform_umax, GroupNonUniformUMax)
+  _SPIRV_OP(group_non_uniform_fmax, GroupNonUniformFMax)
+  _SPIRV_OP(group_non_uniform_iand, GroupNonUniformBitwiseAnd)
+  _SPIRV_OP(group_non_uniform_ior, GroupNonUniformBitwiseOr)
+  _SPIRV_OP(group_non_uniform_ixor, GroupNonUniformBitwiseXor)
+  _SPIRV_OP(group_non_uniform_logical_iand, GroupNonUniformLogicalAnd)
+  _SPIRV_OP(group_non_uniform_logical_ior, GroupNonUniformLogicalOr)
+  _SPIRV_OP(group_non_uniform_logical_ixor, GroupNonUniformLogicalXor)
+  // cl_khr_subgroup_shuffle
+  _SPIRV_OP(group_shuffle, GroupNonUniformShuffle)
+  _SPIRV_OP(group_shuffle_xor, GroupNonUniformShuffleXor)
+  // cl_khr_subgroup_shuffle_relative
+  _SPIRV_OP(group_shuffle_up, GroupNonUniformShuffleUp)
+  _SPIRV_OP(group_shuffle_down, GroupNonUniformShuffleDown)
+#undef _SPIRV_OP
+}
+
+template <> void SPIRVMap<std::string, Op, OCL12Builtin>::init() {
+#define _SPIRV_OP(x, y) add(#x, Op##y);
+  _SPIRV_OP(add, AtomicIAdd)
+  _SPIRV_OP(sub, AtomicISub)
+  _SPIRV_OP(xchg, AtomicExchange)
+  _SPIRV_OP(cmpxchg, AtomicCompareExchange)
+  _SPIRV_OP(inc, AtomicIIncrement)
+  _SPIRV_OP(dec, AtomicIDecrement)
+  _SPIRV_OP(min, AtomicSMin)
+  _SPIRV_OP(max, AtomicSMax)
+  _SPIRV_OP(umin, AtomicUMin)
+  _SPIRV_OP(umax, AtomicUMax)
+  _SPIRV_OP(and, AtomicAnd)
+  _SPIRV_OP(or, AtomicOr)
+  _SPIRV_OP(xor, AtomicXor)
+#undef _SPIRV_OP
+}
+
+// SPV_INTEL_device_side_avc_motion_estimation extension builtins
+class SPIRVSubgroupsAVCIntelInst;
+template <> void SPIRVMap<std::string, Op, SPIRVSubgroupsAVCIntelInst>::init() {
+  // Here is a workaround for a bug in the specification:
+  // 'avc' missed in 'intel_sub_group_avc' prefix.
+  add("intel_sub_group_ime_ref_window_size",
+      OpSubgroupAvcImeRefWindowSizeINTEL);
+
+#define _SPIRV_OP(x, y) add("intel_sub_group_avc_" #x, OpSubgroupAvc##y##INTEL);
+  // Initialization phase functions
+  _SPIRV_OP(ime_initialize, ImeInitialize)
+  _SPIRV_OP(fme_initialize, FmeInitialize)
+  _SPIRV_OP(bme_initialize, BmeInitialize)
+  _SPIRV_OP(sic_initialize, SicInitialize)
+
+  // Result and payload types conversion functions
+  _SPIRV_OP(mce_convert_to_ime_payload, MceConvertToImePayload)
+  _SPIRV_OP(mce_convert_to_ime_result, MceConvertToImeResult)
+  _SPIRV_OP(mce_convert_to_ref_payload, MceConvertToRefPayload)
+  _SPIRV_OP(mce_convert_to_ref_result, MceConvertToRefResult)
+  _SPIRV_OP(mce_convert_to_sic_payload, MceConvertToSicPayload)
+  _SPIRV_OP(mce_convert_to_sic_result, MceConvertToSicResult)
+  _SPIRV_OP(ime_convert_to_mce_payload, ImeConvertToMcePayload)
+  _SPIRV_OP(ime_convert_to_mce_result, ImeConvertToMceResult)
+  _SPIRV_OP(ref_convert_to_mce_payload, RefConvertToMcePayload)
+  _SPIRV_OP(ref_convert_to_mce_result, RefConvertToMceResult)
+  _SPIRV_OP(sic_convert_to_mce_payload, SicConvertToMcePayload)
+  _SPIRV_OP(sic_convert_to_mce_result, SicConvertToMceResult)
+#undef _SPIRV_OP
+
+// MCE instructions
+#define _SPIRV_OP(x, y)                                                        \
+  add("intel_sub_group_avc_mce_" #x, OpSubgroupAvcMce##y##INTEL);
+  _SPIRV_OP(get_default_inter_base_multi_reference_penalty,
+            GetDefaultInterBaseMultiReferencePenalty)
+  _SPIRV_OP(set_inter_base_multi_reference_penalty,
+            SetInterBaseMultiReferencePenalty)
+  _SPIRV_OP(get_default_inter_shape_penalty, GetDefaultInterShapePenalty)
+  _SPIRV_OP(set_inter_shape_penalty, SetInterShapePenalty)
+  _SPIRV_OP(get_default_inter_direction_penalty,
+            GetDefaultInterDirectionPenalty)
+  _SPIRV_OP(set_inter_direction_penalty, SetInterDirectionPenalty)
+  _SPIRV_OP(get_default_intra_luma_shape_penalty,
+            GetDefaultIntraLumaShapePenalty)
+  _SPIRV_OP(get_default_inter_motion_vector_cost_table,
+            GetDefaultInterMotionVectorCostTable)
+  _SPIRV_OP(get_default_high_penalty_cost_table, GetDefaultHighPenaltyCostTable)
+  _SPIRV_OP(get_default_medium_penalty_cost_table,
+            GetDefaultMediumPenaltyCostTable)
+  _SPIRV_OP(get_default_low_penalty_cost_table, GetDefaultLowPenaltyCostTable)
+  _SPIRV_OP(set_motion_vector_cost_function, SetMotionVectorCostFunction)
+  _SPIRV_OP(get_default_intra_luma_mode_penalty, GetDefaultIntraLumaModePenalty)
+  _SPIRV_OP(get_default_non_dc_luma_intra_penalty,
+            GetDefaultNonDcLumaIntraPenalty)
+  _SPIRV_OP(get_default_intra_chroma_mode_base_penalty,
+            GetDefaultIntraChromaModeBasePenalty)
+  _SPIRV_OP(set_ac_only_haar, SetAcOnlyHaar)
+  _SPIRV_OP(set_source_interlaced_field_polarity,
+            SetSourceInterlacedFieldPolarity)
+  _SPIRV_OP(set_single_reference_interlaced_field_polarity,
+            SetSingleReferenceInterlacedFieldPolarity)
+  _SPIRV_OP(set_dual_reference_interlaced_field_polarities,
+            SetDualReferenceInterlacedFieldPolarities)
+  _SPIRV_OP(get_motion_vectors, GetMotionVectors)
+  _SPIRV_OP(get_inter_distortions, GetInterDistortions)
+  _SPIRV_OP(get_best_inter_distortion, GetBestInterDistortions)
+  _SPIRV_OP(get_inter_major_shape, GetInterMajorShape)
+  _SPIRV_OP(get_inter_minor_shapes, GetInterMinorShape)
+  _SPIRV_OP(get_inter_directions, GetInterDirections)
+  _SPIRV_OP(get_inter_motion_vector_count, GetInterMotionVectorCount)
+  _SPIRV_OP(get_inter_reference_ids, GetInterReferenceIds)
+  _SPIRV_OP(get_inter_reference_interlaced_field_polarities,
+            GetInterReferenceInterlacedFieldPolarities)
+#undef _SPIRV_OP
+
+// IME instructions
+#define _SPIRV_OP(x, y)                                                        \
+  add("intel_sub_group_avc_ime_" #x, OpSubgroupAvcIme##y##INTEL);
+  _SPIRV_OP(set_single_reference, SetSingleReference)
+  _SPIRV_OP(set_dual_reference, SetDualReference)
+  _SPIRV_OP(ref_window_size, RefWindowSize)
+  _SPIRV_OP(adjust_ref_offset, AdjustRefOffset)
+  _SPIRV_OP(set_max_motion_vector_count, SetMaxMotionVectorCount)
+  _SPIRV_OP(set_unidirectional_mix_disable, SetUnidirectionalMixDisable)
+  _SPIRV_OP(set_early_search_termination_threshold,
+            SetEarlySearchTerminationThreshold)
+  _SPIRV_OP(set_weighted_sad, SetWeightedSad)
+  _SPIRV_OP(evaluate_with_single_reference, EvaluateWithSingleReference)
+  _SPIRV_OP(evaluate_with_dual_reference, EvaluateWithDualReference)
+  _SPIRV_OP(evaluate_with_single_reference_streamin,
+            EvaluateWithSingleReferenceStreamin)
+  _SPIRV_OP(evaluate_with_dual_reference_streamin,
+            EvaluateWithDualReferenceStreamin)
+  _SPIRV_OP(evaluate_with_single_reference_streamout,
+            EvaluateWithSingleReferenceStreamout)
+  _SPIRV_OP(evaluate_with_dual_reference_streamout,
+            EvaluateWithDualReferenceStreamout)
+  _SPIRV_OP(evaluate_with_single_reference_streaminout,
+            EvaluateWithSingleReferenceStreaminout)
+  _SPIRV_OP(evaluate_with_dual_reference_streaminout,
+            EvaluateWithDualReferenceStreaminout)
+  _SPIRV_OP(get_single_reference_streamin, GetSingleReferenceStreamin)
+  _SPIRV_OP(get_dual_reference_streamin, GetDualReferenceStreamin)
+  _SPIRV_OP(strip_single_reference_streamout, StripSingleReferenceStreamout)
+  _SPIRV_OP(strip_dual_reference_streamout, StripDualReferenceStreamout)
+  _SPIRV_OP(get_border_reached, GetBorderReached)
+  _SPIRV_OP(get_truncated_search_indication, GetTruncatedSearchIndication)
+  _SPIRV_OP(get_unidirectional_early_search_termination,
+            GetUnidirectionalEarlySearchTermination)
+  _SPIRV_OP(get_weighting_pattern_minimum_motion_vector,
+            GetWeightingPatternMinimumMotionVector)
+  _SPIRV_OP(get_weighting_pattern_minimum_distortion,
+            GetWeightingPatternMinimumDistortion)
+#undef _SPIRV_OP
+
+#define _SPIRV_OP(x, y)                                                        \
+  add("intel_sub_group_avc_ime_get_streamout_major_shape_" #x,                 \
+      OpSubgroupAvcImeGetStreamout##y##INTEL);
+  _SPIRV_OP(motion_vectors_single_reference,
+            SingleReferenceMajorShapeMotionVectors)
+  _SPIRV_OP(distortions_single_reference, SingleReferenceMajorShapeDistortions)
+  _SPIRV_OP(reference_ids_single_reference,
+            SingleReferenceMajorShapeReferenceIds)
+  _SPIRV_OP(motion_vectors_dual_reference, DualReferenceMajorShapeMotionVectors)
+  _SPIRV_OP(distortions_dual_reference, DualReferenceMajorShapeDistortions)
+  _SPIRV_OP(reference_ids_dual_reference, DualReferenceMajorShapeReferenceIds)
+#undef _SPIRV_OP
+
+// REF instructions
+#define _SPIRV_OP(x, y)                                                        \
+  add("intel_sub_group_avc_ref_" #x, OpSubgroupAvcRef##y##INTEL);
+  _SPIRV_OP(set_bidirectional_mix_disable, SetBidirectionalMixDisable)
+  _SPIRV_OP(set_bilinear_filter_enable, SetBilinearFilterEnable)
+  _SPIRV_OP(evaluate_with_single_reference, EvaluateWithSingleReference)
+  _SPIRV_OP(evaluate_with_dual_reference, EvaluateWithDualReference)
+  _SPIRV_OP(evaluate_with_multi_reference, EvaluateWithMultiReference)
+  _SPIRV_OP(evaluate_with_multi_reference_interlaced,
+            EvaluateWithMultiReferenceInterlaced)
+#undef _SPIRV_OP
+
+// SIC instructions
+#define _SPIRV_OP(x, y)                                                        \
+  add("intel_sub_group_avc_sic_" #x, OpSubgroupAvcSic##y##INTEL);
+  _SPIRV_OP(configure_skc, ConfigureSkc)
+  _SPIRV_OP(configure_ipe_luma, ConfigureIpeLuma)
+  _SPIRV_OP(configure_ipe_luma_chroma, ConfigureIpeLumaChroma)
+  _SPIRV_OP(get_motion_vector_mask, GetMotionVectorMask)
+  _SPIRV_OP(set_intra_luma_shape_penalty, SetIntraLumaShapePenalty)
+  _SPIRV_OP(set_intra_luma_mode_cost_function, SetIntraLumaModeCostFunction)
+  _SPIRV_OP(set_intra_chroma_mode_cost_function, SetIntraChromaModeCostFunction)
+  _SPIRV_OP(set_skc_bilinear_filter_enable, SetBilinearFilterEnable)
+  _SPIRV_OP(set_skc_forward_transform_enable, SetSkcForwardTransformEnable)
+  _SPIRV_OP(set_block_based_raw_skip_sad, SetBlockBasedRawSkipSad)
+  _SPIRV_OP(evaluate_ipe, EvaluateIpe)
+  _SPIRV_OP(evaluate_with_single_reference, EvaluateWithSingleReference)
+  _SPIRV_OP(evaluate_with_dual_reference, EvaluateWithDualReference)
+  _SPIRV_OP(evaluate_with_multi_reference, EvaluateWithMultiReference)
+  _SPIRV_OP(evaluate_with_multi_reference_interlaced,
+            EvaluateWithMultiReferenceInterlaced)
+  _SPIRV_OP(get_ipe_luma_shape, GetIpeLumaShape)
+  _SPIRV_OP(get_best_ipe_luma_distortion, GetBestIpeLumaDistortion)
+  _SPIRV_OP(get_best_ipe_chroma_distortion, GetBestIpeChromaDistortion)
+  _SPIRV_OP(get_packed_ipe_luma_modes, GetPackedIpeLumaModes)
+  _SPIRV_OP(get_ipe_chroma_mode, GetIpeChromaMode)
+  _SPIRV_OP(get_packed_skc_luma_count_threshold, GetPackedSkcLumaCountThreshold)
+  _SPIRV_OP(get_packed_skc_luma_sum_threshold, GetPackedSkcLumaSumThreshold)
+  _SPIRV_OP(get_inter_raw_sads, GetInterRawSads)
+#undef _SPIRV_OP
+}
+
+template <> void SPIRVMap<std::string, Op, OCLOpaqueType>::init() {
+  add("opencl.event_t", OpTypeEvent);
+  add("opencl.pipe_t", OpTypePipe);
+  add("opencl.clk_event_t", OpTypeDeviceEvent);
+  add("opencl.reserve_id_t", OpTypeReserveId);
+  add("opencl.queue_t", OpTypeQueue);
+  add("opencl.sampler_t", OpTypeSampler);
+}
+
+template <> void LLVMSPIRVAtomicRmwOpCodeMap::init() {
+  add(llvm::AtomicRMWInst::Xchg, OpAtomicExchange);
+  add(llvm::AtomicRMWInst::Add, OpAtomicIAdd);
+  add(llvm::AtomicRMWInst::Sub, OpAtomicISub);
+  add(llvm::AtomicRMWInst::And, OpAtomicAnd);
+  add(llvm::AtomicRMWInst::Or, OpAtomicOr);
+  add(llvm::AtomicRMWInst::Xor, OpAtomicXor);
+  add(llvm::AtomicRMWInst::Max, OpAtomicSMax);
+  add(llvm::AtomicRMWInst::Min, OpAtomicSMin);
+  add(llvm::AtomicRMWInst::UMax, OpAtomicUMax);
+  add(llvm::AtomicRMWInst::UMin, OpAtomicUMin);
+}
+
+} // namespace SPIRV
+
 ///////////////////////////////////////////////////////////////////////////////
 //
 // Functions for getting builtin call info
 //
 ///////////////////////////////////////////////////////////////////////////////
+
+namespace OCLUtil {
+
 AtomicWorkItemFenceLiterals getAtomicWorkItemFenceLiterals(CallInst *CI) {
   return std::make_tuple(getArgAsInt(CI, 0),
                          static_cast<OCLMemOrderKind>(getArgAsInt(CI, 1)),
